backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
anthropic>=0.80.0
python-dotenv==1.0.1
Pillow==11.1.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
numpy>=1.24.0
